:recycle: Score를 다른 패키지로 이동 후 Score를 사용하게 변경
