+++ src/main/java/baseball/gamePlay/Controller.java
@@ -1,6 +1,7 @@
 package baseball.gamePlay;
 
 import baseball.gameStrategy.Balls;
+import baseball.gameStrategy.Score;
 import java.util.List;
 
 // TODO
