+++ src/main/java/baseball/gameStrategy/Match.java
@@ -3,11 +3,11 @@ package baseball.gameStrategy;
 enum Match {
     STRIKE, BALL, NOTHING;
 
-    public boolean isStrike() {
+    boolean isStrike() {
         return this == STRIKE;
     }
 
-    public boolean isBall() {
+    boolean isBall() {
         return this == BALL;
     }
 }
