+++ src/test/java/baseball/gamePlay/ControllerTest.java
@@ -23,9 +23,6 @@ class ControllerTest {
 
     private final List<Integer> allStrikeBalls = Arrays.asList(1, 2, 3);
     private final List<Integer> nothingBalls = Arrays.asList(7, 8, 9);
-    private final Score allStrikeScore = new Score(3, 0);
-    private final Score nothingScore = new Score(0, 0);
-
 
     @DisplayName("플레이어에게 값을 입력받아 컴퓨터의 값과 비교한 결과를 사용자에게 출력한다")
     @Test
@@ -34,7 +31,6 @@ class ControllerTest {
         given(computer.ballNumbers()).willReturn(allStrikeBalls);
         given(player.guessBalls()).willReturn(allStrikeBalls);
         given(player.guessStartNewGame()).willReturn(false);
-//        given(umpire.call(any(), any())).willReturn(allStrikeScore);
         InOrder inOrder = inOrder(computer, player);
 
         //when
@@ -54,7 +50,6 @@ class ControllerTest {
 
         given(computer.ballNumbers()).willReturn(allStrikeBalls);
         given(player.guessBalls()).willReturn(nothingBalls, allStrikeBalls);
-//        given(umpire.call(any(), any())).willReturn(nothingScore, allStrikeScore);
 
         //when
         controller.start();
@@ -74,7 +69,6 @@ class ControllerTest {
         given(computer.ballNumbers()).willReturn(allStrikeBalls);
         given(player.guessBalls()).willReturn(allStrikeBalls);
         given(player.guessStartNewGame()).willReturn(true, false);
-//        given(umpire.call(any(), any())).willReturn(allStrikeScore);
 
         //when
         controller.start();
