+++ src/test/java/baseball/gameStrategy/BallsTest.java
@@ -3,7 +3,6 @@ package baseball.gameStrategy;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
 
-import baseball.gamePlay.Score;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
